% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsMD_helpers.R
\name{rmsMD_calculate_coef_confidence_intervals}
\alias{rmsMD_calculate_coef_confidence_intervals}
\title{Calculating Coefficient Confidence Intervals}
\usage{
rmsMD_calculate_coef_confidence_intervals(output_df)
}
\arguments{
\item{output_df}{A data frame containing \code{coef} and \code{SE} columns.}
}
\value{
A data frame with additional \code{coef_lower95} and \code{coef_upper95} columns.
}
\description{
Adds 95\% confidence intervals for coefficients to the output data frame.
}
\keyword{internal}
