% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsMD_helpers.R
\name{rmsMD_calculate_raw_p_values}
\alias{rmsMD_calculate_raw_p_values}
\title{Calculating Raw p-Values}
\usage{
rmsMD_calculate_raw_p_values(output_df)
}
\arguments{
\item{output_df}{A data frame containing \code{coef} and \code{SE} columns.}
}
\value{
A data frame with an additional \code{p_values_raw} column containing raw p-values.
}
\description{
Calculates p-values from coefficients and standard errors.
}
\keyword{internal}
