% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsMD_helpers.R
\name{rmsMD_combine_CI_for_output}
\alias{rmsMD_combine_CI_for_output}
\title{Combining Confidence Intervals for Output}
\usage{
rmsMD_combine_CI_for_output(output_df, key_vars)
}
\arguments{
\item{output_df}{A data frame containing coefficient, lower, and upper confidence intervals.}

\item{key_vars}{A character vector containing column names for the effect estimate, lower, and upper confidence intervals.}
}
\value{
A data frame with a new column containing combined confidence intervals.
}
\description{
Combines lower and upper confidence intervals into a single formatted string.
}
\keyword{internal}
