% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsMD_helpers.R
\name{rmsMD_extract_coef_and_se}
\alias{rmsMD_extract_coef_and_se}
\title{Extracting Coefficients and Standard Errors}
\usage{
rmsMD_extract_coef_and_se(modelfit)
}
\arguments{
\item{modelfit}{}
}
\value{
two vectors: coef and se
}
\description{
takes model fit, returns two vectors: coef and se
}
\keyword{internal}
