% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsMD_helpers.R
\name{rmsMD_extract_rcs_variables}
\alias{rmsMD_extract_rcs_variables}
\title{Extracting Restricted Cubic Spline (RCS) Variables}
\usage{
rmsMD_extract_rcs_variables(modelfit)
}
\arguments{
\item{modelfit}{A fitted model object}
}
\value{
A list containing two vectors: \code{variables} (RCS variable names) and \code{knots} (number of knots).
}
\description{
Identifies the variables associated with restricted cubic spline (RCS) terms from a model fit object.
}
\keyword{internal}
