% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsMD_helpers.R
\name{rmsMD_format_column_output}
\alias{rmsMD_format_column_output}
\title{Formatting Column For Output}
\usage{
rmsMD_format_column_output(output_df, vars = c(), round_dp_coef)
}
\arguments{
\item{output_df}{A data frame containing the output of model results.}

\item{vars}{A character vector of column names to format.}

\item{round_dp_coef}{The number of decimal places to round the values.}
}
\value{
A data frame with the specified columns formatted.
}
\description{
Formats specific columns in the output data frame by rounding numeric values.
}
\keyword{internal}
