% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsMD_helpers.R
\name{rmsMD_format_final_output}
\alias{rmsMD_format_final_output}
\title{Formatting Final Output}
\usage{
rmsMD_format_final_output(output_df, fullmodel, combine_ci, exp_coef, key_vars)
}
\arguments{
\item{output_df}{A data frame containing model results.}

\item{fullmodel}{A logical indicating whether to return the full model output.}

\item{combine_ci}{A logical indicating whether to combine confidence intervals.}

\item{exp_coef}{A logical indicating whether to use exponentiated coefficients.}

\item{key_vars}{A character vector of key column names to include in the output.}
}
\value{
A formatted data frame for final output.
}
\description{
Formats the final output data frame, including the option to exclude intercept and select columns.
}
\keyword{internal}
