% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsMD_helpers.R
\name{rmsMD_prepare_output_dataframe}
\alias{rmsMD_prepare_output_dataframe}
\title{Preparing the Output Data Frame}
\usage{
rmsMD_prepare_output_dataframe(coef_values, se_values)
}
\arguments{
\item{coef_values}{A named vector of coefficients.}

\item{se_values}{A named vector of standard errors.}
}
\value{
A data frame with variables: \code{variable}, \code{coef}, and \code{SE}.
}
\description{
Creates a data frame with variable names, coefficients, and standard errors.
}
\keyword{internal}
