% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsMD_helpers.R
\name{rmsMD_remove_rcs_coefficients}
\alias{rmsMD_remove_rcs_coefficients}
\title{Removing Restricted Cubic Spline (RCS) Coefficients}
\usage{
rmsMD_remove_rcs_coefficients(coef_values, se_values, rcs_variables, num_knots)
}
\arguments{
\item{coef_values}{A named vector of coefficients.}

\item{se_values}{A named vector of standard errors.}

\item{rcs_variables}{A character vector of RCS variable names.}

\item{num_knots}{A numeric vector indicating the number of knots for each RCS variable.}
}
\value{
A list containing filtered \code{coef} and \code{se} vectors with RCS-related terms removed.
}
\description{
Removes the coefficients and standard errors associated with RCS terms from a model.
}
\keyword{internal}
