% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarymodel_ols_test.R
\name{summarymodel_ols_test}
\alias{summarymodel_ols_test}
\title{Title of the function}
\usage{
summarymodel_ols_test(arg)
}
\arguments{
\item{arg}{blah}
}
\value{
some stuff
}
\description{
this is the short description
}
\examples{
summarymodel_ols_test("eg")
}
