% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsummary_rms.R
\name{modelsummary_rms}
\alias{modelsummary_rms}
\title{Create model summary for rms models}
\usage{
modelsummary_rms(
  modelfit,
  combine_ci = TRUE,
  round_dp_coef = 3,
  round_dp_p = 3,
  rcs_overallp = FALSE,
  hide_rcs_coef = FALSE,
  exp_coef = NULL,
  fullmodel = FALSE
)
}
\arguments{
\item{modelfit}{The output from an rms model.}

\item{combine_ci}{TRUE to combine the effect estimates and 95\% confidence intervals into a single column.}

\item{round_dp_coef}{Specify the number of decimal places to show for the effect estimates.}

\item{round_dp_p}{Specify the number of decimal places to show for P values.}

\item{rcs_overallp}{Set TRUE to give an overall P value for RCS terms (taken from anova(modelfit))}

\item{hide_rcs_coef}{Set TRUE to hide the individual coefficients for RCS variables.}

\item{exp_coef}{Set TRUE to output exp(coef) as the effect estimate (only for model types other than ols, lrm or cph)}

\item{fullmodel}{Set to TRUE to see all of the intermediate steps and check against standard model outputs.}
}
\value{
Returns a dataframe of results. This can easily be outputted to word using
packages such as flextable and officer.
}
\description{
Takes an rms model object and outputs the results in a dataframe.
Outputs results in a format for publication in medical journals.
}
