% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsMD_helpers.R
\name{simulated_rmsMD_data}
\alias{simulated_rmsMD_data}
\title{Simulated data for the Vignette}
\usage{
simulated_rmsMD_data()
}
\description{
Formats the final output data frame, including the option to exclude intercept and select columns.
}
